mcp_servers/word/mcp_servers/docs_server/tools/apply_formatting.py
import os

from docx import Document
from docx.shared import Pt, RGBColor
from helpers.identifier import parse_identifier
from helpers.mutate import resolve_target
from mcp_schema import FlatBaseModel
from pydantic import Field
from utils.decorators import make_async_background
from utils.models import ApplyFormattingResponse, TargetInfo
from utils.path_utils import resolve_under_root


def _parse_color(value: str) -> RGBColor:
    s = value.strip().lstrip("#").upper()
    if len(s) != 6:
        raise ValueError(
            "font_color must be a 6-hex RGB string like 'FF0000' or '#FF0000'"
        )
    r = int(s[0:2], 16)
    g = int(s[2:4], 16)
    b = int(s[4:6], 16)
    return RGBColor(r, g, b)


class ApplyFormattingInput(FlatBaseModel):
    file_path: str = Field(
        ...,
        description="Absolute path to the .docx file starting with '/' (e.g., '/documents/report.docx')",
    )
    identifier: str = Field(
        ...,
        description="Stable identifier from read_document_content for element to format (e.g., 'body.p.0' applies to all runs in paragraph)",
    )
    bold: bool | None = Field(
        None,
        description="Set bold formatting: true to enable, false to disable, null/omit to leave unchanged",
    )
    italic: bool | None = Field(
        None,
        description="Set italic formatting: true to enable, false to disable, null/omit to leave unchanged",
    )
    underline: bool | None = Field(
        None,
        description="Set underline formatting: true to enable, false to disable, null/omit to leave unchanged",
    )
    strikethrough: bool | None = Field(
        None,
        description="Set strikethrough formatting: true to enable, false to disable, null/omit to leave unchanged",
    )
    font_size: float | int | None = Field(
        None,
        description="Font size in points (e.g., 12, 14.5); null/omit to leave unchanged",
    )
    font_color: str | None = Field(
        None,
        description="Font color as 6-character hex RGB string (e.g., 'FF0000' for red, '0000FF' for blue); with or without '#' prefix",
    )


@make_async_background
def apply_formatting(input: ApplyFormattingInput) -> str:
    """Apply text styling attributes to a run, paragraph, or cell selected by identifier."""
    file_path = input.file_path
    identifier = input.identifier
    bold = input.bold
    italic = input.italic
    underline = input.underline
    strikethrough = input.strikethrough
    font_size = input.font_size
    font_color = input.font_color

    if not isinstance(file_path, str) or not file_path:
        return "File path is required"
    if not file_path.startswith("/"):
        return "File path must start with /"
    if not file_path.lower().endswith(".docx"):
        return "File path must end with .docx"
    if not identifier or not identifier.strip():
        return "Identifier is required"

    target_path = resolve_under_root(file_path)

    try:
        if not os.path.exists(target_path):
            return f"File not found: {file_path}"
        if not os.path.isfile(target_path):
            return f"Not a file: {file_path}"

        doc = Document(target_path)
    except Exception as exc:
        return f"Failed to read document: {repr(exc)}"

    # Resolve identifier and target object
    try:
        parsed = parse_identifier(identifier)
        target_kind, target_obj, target_type = resolve_target(doc, parsed)
    except Exception as exc:
        return f"Failed to parse identifier '{identifier}'. Please ensure the identifier is valid from read_document_content tool. Error: {repr(exc)}"

    # Collect runs to apply formatting to
    runs_to_update: list = []
    if target_type == "run":
        runs_to_update = [target_obj]
    elif target_type == "paragraph":
        runs_to_update = list(target_obj.runs)
        if not runs_to_update:
            # ensure there is a run to carry formatting
            target_obj.add_run("")
            runs_to_update = list(target_obj.runs)
    elif target_type == "cell":
        for p in target_obj.paragraphs:
            if not p.runs:
                p.add_run("")
            runs_to_update.extend(list(p.runs))
    else:
        return (
            "Unsupported target for formatting; use run, paragraph, or cell identifier"
        )

    # Apply formatting
    updates_summary: dict = {}

    if bold is not None:
        for r in runs_to_update:
            r.bold = bool(bold)
        updates_summary["bold"] = bool(bold)

    if italic is not None:
        for r in runs_to_update:
            r.italic = bool(italic)
        updates_summary["italic"] = bool(italic)

    if underline is not None:
        for r in runs_to_update:
            r.underline = bool(underline)
        updates_summary["underline"] = bool(underline)

    if strikethrough is not None:
        for r in runs_to_update:
            r.font.strike = bool(strikethrough)
        updates_summary["strikethrough"] = bool(strikethrough)

    if font_size is not None:
        size_pt = float(font_size)
        for r in runs_to_update:
            r.font.size = Pt(size_pt)
        updates_summary["font_size"] = size_pt

    if font_color is not None:
        try:
            color_rgb = _parse_color(font_color)
            for r in runs_to_update:
                r.font.color.rgb = color_rgb
            updates_summary["font_color"] = font_color
        except ValueError as e:
            return str(e)

    # Save document
    try:
        doc.save(target_path)
    except Exception as exc:
        return f"Failed to save document: {repr(exc)}"

    result = ApplyFormattingResponse(
        filepath=file_path,
        status="success",
        target=TargetInfo(kind=target_kind, identifier=identifier),
        applied=updates_summary,
        updated_runs_count=len(runs_to_update),
    )

    return str(result)
