mcp_servers/word/pyproject.toml
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.coverage.run]
source = ["mcp_servers/docs_server/tools"]
omit = ["*/tests/*", "*/__pycache__/*", "*/_meta_tools.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "pass",
]

[project]
name = "mercor-rls-docs"
version = "0.1.0"
description = "Mercor RLS Docs MCP"
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "aioboto3>=15.4.0",
    "aiobotocore>=2.25.0",
    "asyncer>=0.0.9",
    "asyncpg>=0.30.0",
    "datadog-api-client>=2.44.0",
    "fastmcp>=2.12.4",
    "httpx>=0.27.0",
    "litellm>=1.77.7",
    "loguru>=0.7.3",
    "mcp-schema",
    "pydantic-settings>=2.11.0",
    "redis>=6.4.0",
    "python-docx>=1.1.2",
    "pillow>=10.0.0",
]

[dependency-groups]
dev = [
    "basedpyright>=1.31.6",
    "ruff>=0.14.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.2",
    "pytest-cov>=7.0.0",
]

[tool.basedpyright]
typeCheckingMode = "recommended"
reportUnusedCallResult = "none"
failOnWarnings = false
# Style preferences (not errors)
reportMissingTypeArgument = "none"
reportImplicitRelativeImport = "none"
# Pydantic false positives
reportCallIssue = "warning"
exclude = ["scripts", 
    "packages",
    "modal_labs.py",
    "modal_helpers.py",
    ".venv",
    ".git",
    ".tmp",
    "venv",
    "mcp_servers",
    "tests",
    "smoke_test",
]

[tool.ruff.lint]
# Enable linting rules
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "W",   # pycodestyle warnings
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]

ignore = [
    "E501",  # Line too long
]
# Allow long lines in meta-tools files for detailed field descriptions
per-file-ignores = { "mcp_servers/**/tools/_meta_tools.py" = ["E501"] }

[tool.uv.sources]
fastmcp = { git = "https://github.com/jlowin/fastmcp.git", rev = "bc2f601" }
mcp-schema = { path = "packages/mcp_schema" }
